Shift to using graph + op model for the copy_project pipeline
