+++ .gitignore
@@ -13,3 +13,4 @@ __pycache__
 .DS_Store
 .mypy_cache
 .pytest_cache
+dagster_home/
