+++ orchestration/dagster_orchestration/hca_orchestration/tests/pipelines/test_pipelines.py
@@ -12,7 +12,7 @@ from hca_orchestration.pipelines import load_hca, validate_egress
 
 def config_path(relative_path: str) -> str:
     path: str = file_relative_path(
-        __file__, os.path.join("./environments/", relative_path)
+        __file__, os.path.join("../environments/", relative_path)
     )
     return path
