+++ orchestration/dagster_orchestration/pytest.ini
@@ -10,3 +10,6 @@ addopts = -m "not e2e"
 # capture log output (helpful for the e2e tests to monitor progress)
 log_cli = True
 log_level = INFO
+
+
+filterwarnings = ignore::dagster.ExperimentalWarning
